moved container files of deployment center react to their own folder
