+++ client-react/src/pages/app/deployment-center/container/DeploymentCenterContainerLogs.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { ProgressIndicator } from 'office-ui-fabric-react';
 import { useTranslation } from 'react-i18next';
-import { deploymentCenterContent, containerLogs } from './DeploymentCenter.styles';
-import { DeploymentCenterContainerLogsProps } from './DeploymentCenter.types';
+import { deploymentCenterContent, containerLogs } from '../DeploymentCenter.styles';
+import { DeploymentCenterContainerLogsProps } from '../DeploymentCenter.types';
 
 const DeploymentCenterContainerLogs: React.FC<DeploymentCenterContainerLogsProps> = props => {
   const { logs } = props;
