+++ client-react/src/pages/app/deployment-center/container/DeploymentCenterContainerPrivateRegistrySettings.tsx
