+++ client-react/src/pages/app/deployment-center/container/DeploymentCenterContainerRegistrySettings.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import { DeploymentCenterFieldProps, ContainerOptions, ContainerRegistrySources } from './DeploymentCenter.types';
+import { DeploymentCenterFieldProps, ContainerOptions, ContainerRegistrySources } from '../DeploymentCenter.types';
 import { useTranslation } from 'react-i18next';
 import { IChoiceGroupOptionProps } from 'office-ui-fabric-react';
 import { Field } from 'formik';
-import Dropdown from '../../../components/form-controls/DropDown';
+import Dropdown from '../../../../components/form-controls/DropDown';
 
 const DeploymentCenterContainerRegistrySettings: React.FC<DeploymentCenterFieldProps> = props => {
   const { formProps } = props;
